fix: duolingo_user_id BigInteger, case-insensitive username lookup
- User.duolingo_user_id: Integer -> BigInteger (fix integer out of range)
- Migration: users.duolingo_user_id INTEGER -> BIGINT
- get_user_with_languages: case-insensitive username match
- contracts and docker-compose updates

Made-with: Cursor

+++ backend/app/db/models/user.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Boolean, DateTime, Integer, String
+from sqlalchemy import BigInteger, Boolean, DateTime, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base import Base
@@ -10,7 +10,7 @@ class User(Base):
     __tablename__ = "users"
 
     id: Mapped[int] = mapped_column(primary_key=True)
-    duolingo_user_id: Mapped[int | None] = mapped_column(Integer, unique=True, index=True, nullable=True)
+    duolingo_user_id: Mapped[int | None] = mapped_column(BigInteger, unique=True, index=True, nullable=True)
     username: Mapped[str] = mapped_column(String(100), unique=True, index=True)
     display_name: Mapped[str | None] = mapped_column(String(255), nullable=True)
     avatar_url: Mapped[str | None] = mapped_column(String(500), nullable=True)
